quantum_object_detection/requirements.txt
numpy>=1.20.0
pillow>=8.0.0
matplotlib>=3.3.0
torch>=1.8.0
torchvision>=0.9.0
qiskit>=0.25.0
pennylane>=0.15.0
opencv-python>=4.5.0
ultralytics>=8.0.0  # For YOLO implementation
